Cache URI resolving with common doc-cache
To prevent backward-incompatible changes, resolve_uri function
needs to retain lru_cache decorator-compatible function signature
for cache clearing. This change reintroduces the decorator.

The actual data to cache are the documents:
The documents are now read with a separate lru-cached function
and resolve_uri* functions have a customized decorators that
reset nested caches alonside with the lru_cache of
the decorated function itself.